+++ utils/classJournalHref.ts
@@ -1,7 +1,8 @@
-export function buildClassJournalHref(page: number, keyword: string) {
+export function buildClassJournalHref(page: number, keyword = "", mine = false) {
   const params = new URLSearchParams();
   if (page > 1) params.set("page", String(page));
   if (keyword.trim()) params.set("keyword", keyword.trim());
+  if (mine) params.set("mine", "1");
   const query = params.toString();
   return query ? `/staff/class-management?${query}` : "/staff/class-management";
 }
